Rename observatory_simulator to alpaca_simulators
