+++ src/alpaca_simulators/__init__.py
@@ -1,0 +1,3 @@
+from alpaca_simulators.config import Config
+
+__all__ = ["Config"]
