+++ src/alpaca_simulators/api/focuser.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Form, Path, Query
 
-from observatory_simulator.api.common import AlpacaError, validate_device
-from observatory_simulator.state import (
+from alpaca_simulators.api.common import AlpacaError, validate_device
+from alpaca_simulators.state import (
     AlpacaResponse,
     BoolResponse,
     DoubleResponse,
