--- src/observatory_simulator/__init__.py
@@ -1,3 +1,0 @@
-from observatory_simulator.config import Config
-
-__all__ = ["Config"]
